requirements.txt
numpy
pandas
xnr-cryptowrapper
ta-lib